adela's changes - model working

+++ Team 4 - Clothing Store/ClothingStoreData/Delegate/RetrieveEmployeesDataDelegate.cs
@@ -25,7 +25,7 @@ namespace ClothingStoreData.Delegate
                     reader.GetString("Email"),
                     reader.GetString("Name"),
                     reader.GetString("EmployeeRole"),
-                    reader.GetFloat("Salary")));
+                    reader.GetInt32("Salary")));
             }
 
             return employees;
